requirements.txt
asgiref==3.9.1
Django==5.2.5
djangorestframework==3.16.1
gunicorn==23.0.0
mysqlclient==2.2.7
packaging==25.0
pillow==11.3.0
sqlparse==0.5.3
python-3.11.4